relay: Implement simple support for stripping prefix.
The relay protocol sends the prefix of messages in a different element. We
didn't have to do this for glowing-bear, but doing it gives a slightly
closer-to-weechat looking experience (nicknames are in a seperate vertical
column).

I'm unsure if this implementation is completely correct: it looks like weechat
still sends nick prefixes (or at least glowing-bear nails them to the nick), but
including them in key_prefix seems to screw up autocomplete so I don't think
it's quite straightforward.

We should also get nick prefixes out of ISUPPORT instead of just hardcoding
them, but that's for another day.